upgraded to SignalR 2.4.1
I diff'd the existing signalR.js file against the original v2.2.1 to see what changes this package made to the file. I then brought over the new 2.4.1 signalR file and applied the same changes. Those changes are:
* windows.jQuery changed to jQueryShim
* swapped order of parameters for $.inArray (this one seems strange but I didn't investigate why, I just copied the change from the earlier version)
* exports at the bottom of the file.

I also updated the readme to give a more full example of using it with Typescript and pointed people to the @types/signalr-no-jquery  package that provides types.